chore: add jsx as jest file extension
